Fixed bug in management of DB and exception errors.

+++ TilesApp/TilesApp/TilesApp/App.xaml.cs
@@ -95,17 +95,27 @@ namespace TilesApp
         protected override void OnSleep()
         {
             base.OnSleep();
-            int count = Database._database.Table<PendingOperation>().Count();
-            for (int i = 0; i < count; i++)
+            if (App.IsConnected)
             {
-                PendingOperation opt = Database.GetFirstOperationInQueue();
-                if (opt != null)
+                try
                 {
-                    KeyValuePair<string, string> isInserted = CosmosDBManager.InsertOneObject(JsonToOperation(opt));
-                    if (isInserted.Key == "Success")
+                    int count = Database._database.Table<PendingOperation>().Count();
+                    for (int i = 0; i < count; i++)
                     {
-                        Database.DeletePendingOperation(opt);
+                        PendingOperation opt = Database.GetFirstOperationInQueue();
+                        if (opt != null)
+                        {
+                            KeyValuePair<string, string> isInserted = CosmosDBManager.InsertOneObject(JsonToOperation(opt));
+                            if (isInserted.Key == "Success")
+                            {
+                                Database.DeletePendingOperation(opt);
+                            }
+                        }
                     }
+                }
+                catch (Exception e) 
+                {
+                    CrossToastPopUp.Current.ShowToastMessage("There was an error uploading operations. Please, restart de app.");
                 }
             }
         }
